pyproject.toml
[project]
name = "bughunter-blackbox"
version = "0.1.0"
description = "Agentic C++ bug finder using LangGraph + MCP + Groq"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-groq>=0.2.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
]

[project.scripts]
bughunter = "bughunter.__main__:main"
